+++ subcommand/controller/command_test.go
@@ -28,7 +28,7 @@ func TestRun_FlagValidation(t *testing.T) {
 		},
 		{
 			flags:  []string{"-webhook-tls-cert-dir", "/foo", "-datacenter", "foo", "-log-level", "invalid"},
-			expErr: `Error parsing -log-level "invalid": unrecognized level: "invalid"`,
+			expErr: `unknown log level "invalid": unrecognized level: "invalid"`,
 		},
 	}