+++ subcommand/create-federation-secret/command.go
@@ -59,6 +59,7 @@ type Command struct {
 	flagResourcePrefix         string
 	flagK8sNamespace           string
 	flagLogLevel               string
+	flagLogJSON                bool
 	flagMeshGatewayServiceName string
 
 	k8sClient    kubernetes.Interface
@@ -89,6 +90,8 @@ func (c *Command) init() {
 	c.flags.StringVar(&c.flagLogLevel, "log-level", "info",
 		"Log verbosity level. Supported values (in order of detail) are \"trace\", "+
 			"\"debug\", \"info\", \"warn\", and \"error\".")
+	c.flags.BoolVar(&c.flagLogJSON, "log-json", false,
+		"Enable or disable JSON output format for logging.")
 
 	c.http = &flags.HTTPFlags{}
 	c.k8s = &flags.K8SFlags{}
@@ -108,7 +111,7 @@ func (c *Command) Run(args []string) int {
 		return 1
 	}
 
-	logger, err := common.Logger(c.flagLogLevel)
+	logger, err := common.Logger(c.flagLogLevel, c.flagLogJSON)
 	if err != nil {
 		c.UI.Error(err.Error())
 		return 1