+++ subcommand/delete-completed-job/command.go
@@ -18,8 +18,6 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 )
 
-const logLevel = "info"
-
 // Command is the command for deleting completed jobs.
 type Command struct {
 	UI cli.Ui
@@ -28,6 +26,8 @@ type Command struct {
 	k8s           *flags.K8SFlags
 	flagNamespace string
 	flagTimeout   string
+	flagLogLevel  string
+	flagLogJSON   bool
 
 	once      sync.Once
 	help      string
@@ -45,6 +45,11 @@ func (c *Command) init() {
 		"Name of Kubernetes namespace where the job is deployed")
 	c.flags.StringVar(&c.flagTimeout, "timeout", "30m",
 		"How long we'll wait for the job to complete before timing out, e.g. 1ms, 2s, 3m")
+	c.flags.StringVar(&c.flagLogLevel, "log-level", "info",
+		"Log verbosity level. Supported values (in order of detail) are \"trace\", "+
+			"\"debug\", \"info\", \"warn\", and \"error\".")
+	c.flags.BoolVar(&c.flagLogJSON, "log-json", false,
+		"Enable or disable JSON output format for logging.")
 	flags.Merge(c.flags, c.k8s.Flags())
 	c.help = flags.Usage(help, c.flags)
 
@@ -96,7 +101,7 @@ func (c *Command) Run(args []string) int {
 		}
 	}
 
-	logger, err := common.Logger(logLevel)
+	logger, err := common.Logger(c.flagLogLevel, c.flagLogJSON)
 	if err != nil {
 		c.UI.Error(err.Error())
 		return 1