+++ subcommand/sync-catalog/command.go
@@ -52,6 +52,7 @@ type Command struct {
 	flagNodePortSyncType      string
 	flagAddK8SNamespaceSuffix bool
 	flagLogLevel              string
+	flagLogJSON               bool
 
 	// Flags to support namespaces
 	flagEnableNamespaces           bool     // Use namespacing on all components
@@ -122,6 +123,8 @@ func (c *Command) init() {
 	c.flags.StringVar(&c.flagLogLevel, "log-level", "info",
 		"Log verbosity level. Supported values (in order of detail) are \"trace\", "+
 			"\"debug\", \"info\", \"warn\", and \"error\".")
+	c.flags.BoolVar(&c.flagLogJSON, "log-json", false,
+		"Enable or disable JSON output format for logging.")
 
 	c.flags.Var((*flags.AppendSliceValue)(&c.flagAllowK8sNamespacesList), "allow-k8s-namespace",
 		"K8s namespaces to explicitly allow. May be specified multiple times.")
@@ -200,7 +203,7 @@ func (c *Command) Run(args []string) int {
 	// Set up logging
 	if c.logger == nil {
 		var err error
-		c.logger, err = common.Logger(c.flagLogLevel)
+		c.logger, err = common.Logger(c.flagLogLevel, c.flagLogJSON)
 		if err != nil {
 			c.UI.Error(err.Error())
 			return 1